Created ExecutableMessageFactory class to generate ExecutableMessages
